We can now register users, but no error handling.
As of now we can add users names and their passwords to our database. If there are any errors in the user entered information the user will not know, because we haven't displayed error messages to them yet.  If their username is not unique the site will not do anything but you will see an error message in the console. Passwords are not hashed yet, so don't save anything besides the word password.